uadk/test: move main function to testsuit.c and rm
test_sva_perf.c

test_suit.c should be the main function file,
and these two files have the duplicate code.
